backend/requirements.txt
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt == 4.1.1
click==8.1.8
email-validator == 2.1.0
fastapi==0.115.12
frozenlist==1.5.0
gremlinpython==3.7.3
h11==0.14.0
httpx==0.25.2
idna==3.10
isodate==0.7.2
multidict==6.2.0
nest-asyncio==1.6.0
Pillow>=10.1.0  # For image resizing and processing
propcache==0.3.1
pydantic==2.11.0
pydantic_core==2.33.0
PyJWT == 2.8.0
python-dotenv==1.0.0
python-multipart>=0.0.6  # For file uploads with FastAPI
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
yarl==1.18.3
